requirements.txt
altgraph==0.17.3
backports.functools-lru-cache==1.6.4
certifi==2021.10.8
chardet==4.0.0
dis3==0.1.3
docopt==0.6.2
future==0.18.3
idna==2.10
pefile==2019.4.18
pip==20.3.4
prompt-toolkit==2.0.10
pyflakes==2.4.0
Pygments==2.5.2
PyInstaller==3.6
pypiwin32==223
pyvim==2.0.21
pyvmomi==6.7.1
pywin32-ctypes==0.2.0
pywin32==228
PyYAML==5.4.1
requests==2.27.1
setuptools==44.1.1
six==1.16.0
urllib3==1.26.15
wcwidth==0.2.6
wheel==0.37.1
